fix: struct name collision with `axum::debug_handler`
It appears that axum's debug handler macro hardcodes certain types as requiring consuming the body. Because this struct is named `Request`, it matches the hardcoded list of request body consumers, which in turn makes the compiler think that the extractor should be the last in the list of handler arguments. When the extractor is moved to the last argument, the compiler complains about missing the `FromRequest` implementation. Just rename the struct to `Query` to solve this issue, which is also more specific to its usecase.