fix: don't mark peers as invisible on 2nd announce if full dl slots
Peers were being marked as invisible if their leech slots were full on subsequent announces as a leech. This freed up a leech slot on their second announce, and the user could continue downloading with cached peers in their client, and start another torrent with the extra available leech slot.