update: assume delta of 0 on new session
The old logic assumed what was provided is the delta even if non-zero in case a peer was blocked, made inactive, deleted, and came back and we needed to log their stats as if they were always there. We now handle this edge case using the \`visible\` flag, so we no longer need this workaround.