update: treat partial seeds as leeches
If a user selects to only leech part of a torrent, then once finished leeching that selected portion, they will only send a `paused` event. This event triggers an error, which the peer then times out and the user gets hnrs. With this change, the event is accepted, but the peers are treated as leeches so that the user is aware the peer isn't treated as a seed, and therefore is aware that the required seedtime hasn't started yet.