fix: only upsert peer update if it's the latest
If an upsert into the MySQL database fails, then the peer update will be upserted back into the queue, but only if it doesn't already exist. On the other hand, if a user announces twice between upserts, then we only want to save the latest peer update. In all cases, we only want to save the latest peer update so that the MySQL peers table is up to date.