fix: synchronize peer count aggregates on import
When peers are deactivated in the tracker due to inactivity, they are not
immediately synced to the database, because UNIT3D handles the responsibility of
deactivating peers older than 2 hours every 2 minutes. If peers are deactivated
in the tracker, and the tracker is restarted before UNIT3D deactivates the peers
in the database, then the database will have correct torrent seeder/leecher
counts but incorrect peer active status. When the tracker is started up again,
the peers that should be inactive are reimported as active, and proceed to be
deactivated a second time on the next peer expiry interval, decrementing the
torrent seeder/leecher count one less than it should.

Fix this by synchronizing the torrent seeder/leecher count before importing the
undeactivated peers. Note: the combination of recently updated peer and inactive
status is a special case to signify the peer sent a stopped event and should be
deleted. This means we can't simply synchronize the peer's active state to the
database on deactivation.