fix: count invisible peers in allowed max peers per torrent per user
Otherwise, if the user has already filled up their download slots, the user can add 100k invisible peers to a single torrent, and the O(n) loop that checks the number of active peers on a torrent will slow down the tracker. We don't have to worry about peer count delta logic here, because we immediately remove the newly added peer and early return, so we are in the exact same state as before the announce and the deltas are not updated.